+++ src/extractors/standards/format-foreign-tax-id20.ts
@@ -2,11 +2,11 @@ import { html_entities } from '../../utils';
 
 export class FormatForeignTaxId20 {
     public formatForeignTaxId(foreignTaxId: string): string {
-        // codificar
+        // Codificar
         foreignTaxId = html_entities(foreignTaxId);
-        // usar hasta un máximo de 20 posiciones
-        foreignTaxId = foreignTaxId.substring(0, 20);
-        // crear un padding para establecer a 20 posiciones
+        // Usar hasta un máximo de 20 posiciones
+        foreignTaxId = foreignTaxId.slice(0, 20);
+        // Crear un padding para establecer a 20 posiciones
 
         return foreignTaxId.padStart(20, '0');
     }