+++ tests/unit/extractors/standards/format-rfc-xml.test.ts
@@ -6,7 +6,7 @@ describe('FormatRfcXml', () => {
         ['AAAA010101AAA', 'AAAA010101AAA'],
         ['ÑAAA010101AAA', 'ÑAAA010101AAA'],
         ['&AAA010101AAA', '&amp;AAA010101AAA']
-    ])('format rfc', (input: string, expected: string) => {
+    ])('format rfc input: %s - expected: %s', (input: string, expected: string) => {
         const extractor = new FormatRfcXml();
 
         expect(extractor.formatRfc(input)).toBe(expected);